Fixed typing error
MSVC++10 refused to compile this line without proper typing.

+++ include/components/GLScreenQuad.h
@@ -26,7 +26,7 @@ public:
 
 		while(dim < width && dim < height) {
 			power += 1;
-			dim = (int)pow(2, power);
+			dim = (int)pow(2.0, (int)power);
 		}
 
 		return dim;
